requirements.txt
Django==1.11.9
django-easy-pdf==0.1.1
html5lib==1.0.1
httplib2==0.11.3
Pillow==5.1.0
pkg-resources==0.0.0
PyPDF2==1.26.0
pytz==2018.4
reportlab==3.4.0
six==1.11.0
webencodings==0.5.1
xhtml2pdf==0.2.2
